+++ pylatex/base_classes/__init__.py
@@ -6,8 +6,9 @@ Baseclasses that can be used to create classes representing LaTeX objects.
 """
 
 from .latex_object import LatexObject
-from .containers import Container, Environment
-from .command import CommandBase, Command, UnsafeCommand, Options, Arguments
+from .containers import Container, Environment, ContainerCommand
+from .command import CommandBase, Command, UnsafeCommand, Options, \
+    SpecialOptions, Arguments
 from .float import Float
 
 # Old names of the base classes for backwards compatibility