+++ pylatex/lists.py
@@ -8,7 +8,9 @@ These lists are specifically enumerate, itemize and description.
     :license: MIT, see License for more details.
 """
 
-from .base_classes import Environment, Command
+from .base_classes import Environment, Command, Options
+from .package import Package
+from pylatex.utils import NoEscape
 
 
 class List(Environment):
@@ -32,6 +34,38 @@ class List(Environment):
 
 class Enumerate(List):
     """A class that represents an enumerate list."""
+
+    omit_if_empty = False
+
+    def __init__(self, enumeration_symbol=None, *, options=None, **kwargs):
+        r"""
+        Args
+        ----
+        enumeration_symbol: str
+            The enumeration symbol to use, see the `enumitem
+            <https://www.ctan.org/pkg/enumitem>`_ documentation to see what
+            can be used here. This argument is not escaped as it usually
+            should usually contain commands, so do not use user input here.
+        options: str or list or `.Options`
+            Custom options to be added to the enumerate list. These options are
+            merged with the options created by ``enumeration_symbol``.
+        """
+
+        self._enumeration_symbol = enumeration_symbol
+
+        if enumeration_symbol is not None:
+            self.packages.add(Package("enumitem"))
+
+            if options is not None:
+                options = Options(options)
+            else:
+                options = Options()
+            options._positional_args.append(NoEscape('label=' +
+                                                     enumeration_symbol))
+
+        super().__init__(options=options, **kwargs)
+
+        print(self.dumps())
 
 
 class Itemize(List):