+++ tests/args.py
@@ -14,15 +14,27 @@ import matplotlib
 
 from pylatex import Document, Section, Math, Tabular, Figure, SubFigure, \
     Package, TikZ, Axis, Plot, Itemize, Enumerate, Description, MultiColumn, \
-    MultiRow, Command, Matrix, VectorName, Quantity, TableRowSizeError
+    MultiRow, Command, Matrix, VectorName, Quantity, TableRowSizeError, \
+    LongTable, FlushLeft, FlushRight, Center, MiniPage, TextBlock, \
+    PageStyle, Head, Foot, StandAloneGraphic, Tabularx, Column, NewLine, \
+    LineBreak, NewPage, HFill, HugeText, LargeText, MediumText, \
+    SmallText, FootnoteText, TextColor, FBox, MdFramed, Tabu, \
+    HorizontalSpace, VerticalSpace
 from pylatex.utils import escape_latex, fix_filename, dumps_list, bold, \
-    italic, verbatim
+    italic, verbatim, NoEscape
 
 matplotlib.use('Agg')  # Not to use X server. For TravisCI.
 import matplotlib.pyplot as pyplot  # noqa
 
 
 def test_document():
+    geometry_options = {
+        "includeheadfoot": True,
+        "headheight": "12pt",
+        "headsep": "10pt",
+        "landscape": True
+    }
+
     doc = Document(
         default_filepath='default_filepath',
         documentclass='article',
@@ -30,11 +42,22 @@ def test_document():
         inputenc='utf8',
         lmodern=True,
         data=None,
+        page_numbers=True,
+        indent=False,
+        document_options=["a4paper", "12pt"],
+        geometry_options=geometry_options
     )
 
     repr(doc)
 
     doc.append('Some text.')
+    doc.change_page_style(style="empty")
+    doc.change_document_style(style="plain")
+    doc.add_color(name="lightgray", model="gray", description="0.6")
+    doc.add_color(name="abitless", model="gray", description="0.8")
+    doc.set_variable(name="myVar", value="1234")
+    doc.set_variable(name="myVar", value="1234")
+    doc.change_length(parameter=r"\headheight", value="0.5in")
 
     doc.generate_tex(filepath='')
     doc.generate_pdf(filepath='', clean=True)
@@ -67,7 +90,7 @@ def test_table():
 
     t.add_hline(start=None, end=None)
 
-    t.add_row(cells=(1, 2), escape=False, strict=True)
+    t.add_row(cells=(1, 2), escape=False, strict=True, mapper=[bold])
 
     # MultiColumn/MultiRow.
     t.add_row((MultiColumn(size=2, align='|c|', data='MultiColumn'),),
@@ -78,6 +101,27 @@ def test_table():
     t.add_row((MultiRow(size=2, width='*', data='MultiRow'),), strict=False)
 
     repr(t)
+
+    # TabularX
+    tabularx = Tabularx(table_spec='X X X', arguments=NoEscape(r"\textwidth"))
+    tabularx.add_row(["test1", "test2", "test3"])
+
+    # Long Table
+    longtable = LongTable(table_spec='c c c')
+    longtable.add_row(["test", "test2", "test3"])
+    longtable.end_table_header()
+
+    # Colored Tabu
+    coloredtable = Tabu(table_spec='X[c] X[c]')
+    coloredtable.add_row(["test", "test2"], color="gray", mapper=bold)
+
+    # Colored Tabularx
+    coloredtable = Tabularx(table_spec='X[c] X[c]')
+    coloredtable.add_row(["test", "test2"], color="gray", mapper=bold)
+
+    # Column
+    column = Column("R", "X", r"\raggedleft", parameters=2)
+    repr(column)
 
 
 def test_command():
@@ -115,6 +159,11 @@ def test_graphics():
     plot.add_caption(caption='I am a caption.')
     repr(plot)
 
+    # StandAloneGraphic
+    stand_alone_graphic = StandAloneGraphic(
+        filename='', image_options=r"width=0.8\textwidth")
+    repr(stand_alone_graphic)
+
 
 def test_quantities():
     # Quantities
@@ -149,8 +198,9 @@ def test_lists():
     itemize.append("append")
     repr(itemize)
 
-    enum = Enumerate()
+    enum = Enumerate(enumeration_symbol=r"\alph*)", options={'start': 172})
     enum.add_item(s="item")
+    enum.add_item(s="item2")
     enum.append("append")
     repr(enum)
 
@@ -158,6 +208,109 @@ def test_lists():
     desc.add_item(label="label", s="item")
     desc.append("append")
     repr(desc)
+
+
+def test_headfoot():
+    # Page styles, headers and footers
+    page_style = PageStyle("NewStyle")
+    page_style.change_thickness("header", "1pt")
+    page_style.change_thickness("footer", "1pt")
+
+    header = Head("C")
+    header.append("append")
+
+    footer = Foot("C")
+    footer.append("append")
+
+    page_style.append(header)
+    page_style.append(footer)
+    repr(header)
+    repr(footer)
+    repr(page_style)
+
+
+def test_position():
+
+    repr(HorizontalSpace(size='20pt', star=False))
+
+    repr(VerticalSpace(size="20pt", star=True))
+
+    # Test alignment environments
+    center = Center()
+    center.append("append")
+    repr(center)
+
+    right = FlushRight()
+    right.append("append")
+    repr(right)
+
+    left = FlushLeft()
+    left.append("append")
+    repr(left)
+
+    minipage = MiniPage(width=r"\textwidth", height="10pt", pos='t',
+                        align='r', content_pos='t', fontsize="Large")
+    minipage.append("append")
+    repr(minipage)
+
+    textblock = TextBlock(width="200", horizontal_pos="200",
+                          vertical_pos="200", indent=True)
+    textblock.append("append")
+    textblock.dumps()
+    repr(textblock)
+
+
+def test_frames():
+    # Tests the framed environments
+
+    md_framed = MdFramed()
+    md_framed.append("Framed text")
+    repr(md_framed)
+
+    f_box = FBox()
+    f_box.append("Fboxed text")
+    repr(f_box)
+
+
+def test_basic():
+    # Tests the basic commands and environments
+    # Basic commands
+    new_page = NewPage()
+    repr(new_page)
+
+    new_line = NewLine()
+    repr(new_line)
+
+    line_break = LineBreak()
+    repr(line_break)
+
+    h_fill = HFill()
+    repr(h_fill)
+
+    # Basic environments
+    huge = HugeText("Huge")
+    huge.append("Huge 2")
+    repr(huge)
+
+    large = LargeText("Large")
+    large.append("Large 2")
+    repr(large)
+
+    medium = MediumText("Medium")
+    medium.append("Medium 2")
+    repr(medium)
+
+    small = SmallText("Small")
+    small.append("Small 2")
+    repr(small)
+
+    footnote = FootnoteText("Footnote")
+    footnote.append("Footnote 2")
+    repr(footnote)
+
+    text_color = TextColor("green", "GreenText")
+    text_color.append("More Green Text")
+    repr(text_color)
 
 
 def test_utils():